`optype infer` CLI entrypoint
